EDU-583 Separate url utils out of routes module
