+++ src/utils/routes.spec.js
@@ -3,48 +3,6 @@ import sut from './routes.js';
 describe('routes', () => {
   let result;
 
-  describe('concatParts', () => {
-    const testCases = [
-      {
-        parts: [null, ''],
-        expectedResult: ''
-      },
-      {
-        parts: ['abc', 'def', 'ghi'],
-        expectedResult: 'abc/def/ghi'
-      },
-      {
-        parts: ['abc', 0, 'ghi'],
-        expectedResult: 'abc/0/ghi'
-      },
-      {
-        parts: ['abc', false, 'ghi'],
-        expectedResult: 'abc/false/ghi'
-      },
-      {
-        parts: ['abc', null, 'ghi'],
-        expectedResult: 'abc/ghi'
-      },
-      {
-        parts: ['abc', '', 'ghi'],
-        expectedResult: 'abc/ghi'
-      }
-    ];
-
-    testCases.forEach(({ parts, expectedResult }) => {
-      describe(`when parts are ${parts}`, () => {
-        let actualResult;
-        beforeEach(() => {
-          actualResult = sut.concatParts(...parts);
-        });
-        it(`should return '${expectedResult}'`, () => {
-          expect(actualResult).toBe(expectedResult);
-        });
-      });
-    });
-
-  });
-
   describe('getDocUrl', () => {
     const testCases = [
       {
