+++ src/utils/url-utils.spec.js
@@ -1,7 +1,4 @@
-import {
-  removeLeadingSlash,
-  removeTrailingSlash
-} from './url-utils.js';
+import sut from './url-utils.js';
 
 describe('url-utils', () => {
   let result;
@@ -25,7 +22,7 @@ describe('url-utils', () => {
     testCases.forEach(({ path, expectedResult }) => {
       describe(`when path is '${path}'`, () => {
         beforeEach(() => {
-          result = removeTrailingSlash(path);
+          result = sut.removeTrailingSlash(path);
         });
         it(`should return '${expectedResult}'`, () => {
           expect(result).toBe(expectedResult);
@@ -55,12 +52,54 @@ describe('url-utils', () => {
     testCases.forEach(({ path, expectedResult }) => {
       describe(`when path is '${path}'`, () => {
         beforeEach(() => {
-          result = removeLeadingSlash(path);
+          result = sut.removeLeadingSlash(path);
         });
         it(`it should return '${expectedResult}'`, () => {
           expect(result).toBe(expectedResult);
         });
       });
     });
+  });
+
+  describe('concatParts', () => {
+    const testCases = [
+      {
+        parts: [null, ''],
+        expectedResult: ''
+      },
+      {
+        parts: ['abc', 'def', 'ghi'],
+        expectedResult: 'abc/def/ghi'
+      },
+      {
+        parts: ['abc', 0, 'ghi'],
+        expectedResult: 'abc/0/ghi'
+      },
+      {
+        parts: ['abc', false, 'ghi'],
+        expectedResult: 'abc/false/ghi'
+      },
+      {
+        parts: ['abc', null, 'ghi'],
+        expectedResult: 'abc/ghi'
+      },
+      {
+        parts: ['abc', '', 'ghi'],
+        expectedResult: 'abc/ghi'
+      }
+    ];
+
+    testCases.forEach(({ parts, expectedResult }) => {
+      describe(`when parts are ${parts}`, () => {
+        let actualResult;
+        beforeEach(() => {
+          actualResult = sut.concatParts(...parts);
+        });
+        it(`should return '${expectedResult}'`, () => {
+          expect(actualResult).toBe(expectedResult);
+        });
+      });
+    });
+
   });
 });
